Rename DapAccess to UntypedDapAccess, combine ApAccess and DpAccess into TypedDapAccess
