--- probe-rs/src/core/communication_interface.rs
@@ -1,5 +1,0 @@
-use crate::DebugProbeError;
-
-pub trait CommunicationInterface {
-    fn flush(&mut self) -> Result<(), DebugProbeError>;
-}
