+++ probe-rs/src/lib.rs
@@ -77,9 +77,8 @@ mod session;
 
 pub use crate::config::{CoreType, Target};
 pub use crate::core::{
-    Architecture, Breakpoint, BreakpointId, CommunicationInterface, Core, CoreInformation,
-    CoreInterface, CoreList, CoreRegister, CoreRegisterAddress, CoreState, CoreStatus, HaltReason,
-    SpecificCoreState,
+    Architecture, Breakpoint, BreakpointId, Core, CoreInformation, CoreInterface, CoreList,
+    CoreRegister, CoreRegisterAddress, CoreState, CoreStatus, HaltReason, SpecificCoreState,
 };
 pub use crate::error::Error;
 pub use crate::memory::{Memory, MemoryInterface};
